@@ -1,6 +1,6 @@
 {
   "name": "template-for-web-app",
-  "version": "0.0.0",
+  "version": "0.0.5",
   "private": true,
   "scripts": {
     "dev": "vite",
@@ -15,39 +15,43 @@
     "format": "prettier --write src/"
   },
   "dependencies": {
-    "@capacitor/android": "^5.4.1",
-    "@capacitor/core": "^5.4.1",
-    "@capacitor/ios": "^5.4.1",
-    "pinia": "^2.1.6",
-    "vue": "^3.3.4",
+    "@capacitor/android": "^5.5.0",
+    "@capacitor/core": "^5.5.0",
+    "@capacitor/ios": "^5.5.0",
+    "@mdi/font": "^7.3.67",
+    "axios": "^1.5.1",
+    "pinia": "^2.1.7",
+    "socket.io-client": "^4.7.2",
+    "vue": "^3.3.6",
     "vue-i18n": "^9.5.0",
     "vue-router": "^4.2.5",
-    "vuetify": "^3.3.19"
+    "vue3-toastify": "^0.1.14",
+    "vuetify": "^3.3.22"
   },
   "devDependencies": {
-    "@capacitor/cli": "^5.4.1",
+    "@capacitor/cli": "^5.5.0",
     "@rushstack/eslint-patch": "^1.5.1",
     "@tsconfig/node18": "^18.2.2",
-    "@types/jsdom": "^21.1.3",
-    "@types/node": "^20.8.2",
+    "@types/jsdom": "^21.1.4",
+    "@types/node": "^20.8.8",
     "@vitejs/plugin-vue": "^4.4.0",
     "@vitejs/plugin-vue-jsx": "^3.0.2",
     "@vue/eslint-config-prettier": "^8.0.0",
     "@vue/eslint-config-typescript": "^12.0.0",
     "@vue/test-utils": "^2.4.1",
     "@vue/tsconfig": "^0.4.0",
-    "cypress": "^13.3.0",
-    "eslint": "^8.50.0",
+    "cypress": "^13.3.3",
+    "eslint": "^8.52.0",
     "eslint-plugin-cypress": "^2.15.1",
-    "eslint-plugin-vue": "^9.17.0",
+    "eslint-plugin-vue": "^9.18.0",
     "jsdom": "^22.1.0",
-    "npm-run-all2": "^6.0.6",
+    "npm-run-all2": "^6.1.1",
     "prettier": "^3.0.3",
     "start-server-and-test": "^2.0.1",
     "typescript": "~5.2.2",
-    "vite": "^4.4.9",
+    "vite": "^4.5.0",
     "vite-plugin-pwa": "^0.16.5",
     "vitest": "^0.34.6",
-    "vue-tsc": "^1.8.15"
+    "vue-tsc": "^1.8.20"
   }
 }
