@@ -3,11 +3,17 @@ import { defineStore } from 'pinia'
 export const useThemeStore = defineStore({
   id: 'theme',
   state: () => ({
-    isLight: false
+    isLight: window.matchMedia('(prefers-color-scheme: light)').matches || false
   }),
   actions: {
+    checkUserTheme() {
+      localStorage.getItem('theme')
+        ? this.$patch({ isLight: true })
+        : this.$patch({ isLight: false })
+    },
     toggleTheme() {
       this.$patch({ isLight: !this.isLight })
+      localStorage.setItem('theme', JSON.stringify(this.isLight))
     }
   }
 })
